fix: move ai prompt settings into assistant tabs
